added downsample to CLI
Former-commit-id: b074fb5b08db092050a8cab93fc78757765c8ce5

+++ aspire/preprocessor/cryo_crop.py
@@ -50,12 +50,12 @@ def cryo_crop(mat, n, stack=False, fill_value=0):
         mat = numpy.reshape(mat, [mat.size, 1])  # force a column vector
         ns = math.floor(mat.size / 2) - math.floor(n / 2)  # shift term for scaling down
         if ns >= 0:  # cropping
-            return mat[ns: ns + n]
+            return mat[ns: ns + n].astype('float32')
 
         else:  # padding
             result_mat = fill_value * numpy.ones([n, 1], dtype=complex)
             result_mat[-ns: mat.size - ns] = mat
-            return result_mat
+            return result_mat.astype('float32')
 
     elif num_dimensions == 2:  # mat is 2D image
         mat_x, mat_y = mat.shape
@@ -67,14 +67,14 @@ def cryo_crop(mat, n, stack=False, fill_value=0):
         if start_x >= 0 and start_y >= 0:  # cropping
             start_x, start_y = math.floor(start_x), math.floor(start_y)
             logger.debug(f'cryo_crop:cropping from {mat.shape} to {n}..')
-            return mat[start_x: start_x + int(n), start_y: start_y + int(n)]
+            return mat[start_x: start_x + int(n), start_y: start_y + int(n)].astype('float32')
 
         elif start_x < 0 and start_y < 0:  # padding
             logger.debug('cryo_crop:padding..')
             start_x, start_y = math.floor(start_x), math.floor(start_y)
             result_mat = fill_value * numpy.ones([n, n], dtype=complex)
             result_mat[-start_x: mat_x - start_x, -start_y: mat_y - start_y] = mat
-            return result_mat
+            return result_mat.astype('float32')
 
         else:
             raise DimensionsIncompatible("Can't crop and pad simultaneously!")
@@ -89,7 +89,7 @@ def cryo_crop(mat, n, stack=False, fill_value=0):
                 # TODO iterate instead of using recursion. this is too memoery-expensive
                 result_mat[img, :, :] = cryo_crop(mat[img, :, :], n)
 
-            return result_mat
+            return result_mat.astype('float32')
 
         else:  # this is a 3D structure
             # crop/pad mat into a new smaller/bigger cell - 'destination cell'
@@ -110,7 +110,7 @@ def cryo_crop(mat, n, stack=False, fill_value=0):
                            -ns[1]: from_shape[2] - ns[1],
                            -ns[2]: from_shape[2] - ns[2]] = mat
 
-                return result_mat
+                return result_mat.astype('float32')
 
             else:
                 raise DimensionsIncompatible("Can't crop and pad simultaneously!")
